Update author name and version in comments throughout the code
Several Java classes have been updated to correct the author name in the comments and update the version number to 1.6.6. This will helps maintain accurate documentation of authors and version histories within the project's codebase in the future.

Another run will be done to credit the correct contributors and version in the future