Rename core.Paragraph to core.DefaultParagraph
The Paragraph class in 'pro.verron.docxstamper.core' package has been renamed to DefaultParagraph and implemented as an interface in 'pro.verron.docxstamper.api' package. All existing references to the Paragraph class in other packages have been updated to reflect this change. As a result, the structure is more flexible for future extensions or alternative implementations of Paragraph handling.