Refactor project structure and move classes for better organization
The changes made to the project mainly involve restructuring and organizing the project for better modularity and organization. Multiple classes have been moved to new packages and some have been renamed to better reflect their functionality. This will improve the maintainability and understanding of the codebase. The enhancement also involves moving the 'DocxStamperConfiguration' and 'DocxStamper' classes to a new package and made them less visible to allow for bigger projects to develop on top of them.