Add office stamper examples
Added several classes in 'officestamper-examples' package that demonstrate the use of the office stamper tool. The pivotal files are 'Examples.java', 'Main.java', and 'Diagnostic.java'. This commit also includes a new 'pom.xml' file for project build configuration and a few helper classes. The '.gitignore' file was updated to properly ignore the '/target/' directory.