Add extensive module information to module-info.java
A comprehensive documentation has been added to the module-info.java file for the 'pro.verron.officestamper' application. This includes declaring its dependencies, packages needed for runtime access and packages that need to be exported. A warning has also been added about certain packages opening and exporting that should be inaccessible in future versions.