Refactor StandardParagraph to enforce R class replacements
The changes are primarily focused on enforcing the replacement parameter to be of class R in the replace method of StandardParagraph. The code also removes the unused 'getRuns' method. The validation is implemented by throwing an AssertionError if the replacement is not an instance of R. This change enforces the stringent requirement while simplifying the overall structure of the code.