[BREAK] Remove deprecated ITypeResolver interface
This commit removes the deprecated ITypeResolver interface and updates the reference of this interface in the AbstractToTextResolver class to use the ObjectResolver from the pro.verron.officestamper.api package. This change is part of an effort to simplify the internal workings of the docx-stamper project and reduce reliance on the reflection mechanism.