Relocate comment simulation to Paragraph class to eliminate getP() method
The comment imitation mechanism, previously located in the CommentProcessors class, has been moved to the Paragraph class. Consequently, the Paragraph interface was updated with a new fakeComment() method, which allows each type of Paragraph to simulate a comment relevant to its specifics. This amendment permitted removal of the getP() method from the Paragraph interface, as it's no longer needed, implementing a cleaner API design.